+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Added
 - Null values are not serialized
 - Program prints version
+- Support to attach examples via json file
 
 ## [0.2.0] - 2021-02-01
 ### Added